x/reward/types/message_claim_reward_request_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/exocore/testutil/sample"
)

func TestMsgClaimRewardRequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgClaimRewardRequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgClaimRewardRequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgClaimRewardRequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
